+++ previous-versions/assess/pkg/models/operations/generateexcelreport.go
@@ -31,8 +31,6 @@ func (o *GenerateExcelReportRequest) GetReportType() shared.ExcelReportType {
 type GenerateExcelReportResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your API request was not properly authorized.
-	ErrorMessage *shared.ErrorMessage
 	// OK
 	ExcelStatus *shared.ExcelStatus
 	// HTTP response status code for this operation
@@ -46,13 +44,6 @@ func (o *GenerateExcelReportResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *GenerateExcelReportResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *GenerateExcelReportResponse) GetExcelStatus() *shared.ExcelStatus {