+++ previous-versions/assess/pkg/models/shared/enhancedreport.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ericlagergren/decimal"
 )
 
-type EnhancedReportReportItem struct {
+type ReportItem struct {
 	AccountCategory *EnhancedReportAccountCategory `json:"accountCategory,omitempty"`
 	// The unique account ID.
 	AccountID *string `json:"accountId,omitempty"`
@@ -37,46 +37,46 @@ type EnhancedReportReportItem struct {
 	Date *string `json:"date,omitempty"`
 }
 
-func (e EnhancedReportReportItem) MarshalJSON() ([]byte, error) {
-	return utils.MarshalJSON(e, "", false)
+func (r ReportItem) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(r, "", false)
 }
 
-func (e *EnhancedReportReportItem) UnmarshalJSON(data []byte) error {
-	if err := utils.UnmarshalJSON(data, &e, "", false, false); err != nil {
+func (r *ReportItem) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &r, "", false, false); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (o *EnhancedReportReportItem) GetAccountCategory() *EnhancedReportAccountCategory {
+func (o *ReportItem) GetAccountCategory() *EnhancedReportAccountCategory {
 	if o == nil {
 		return nil
 	}
 	return o.AccountCategory
 }
 
-func (o *EnhancedReportReportItem) GetAccountID() *string {
+func (o *ReportItem) GetAccountID() *string {
 	if o == nil {
 		return nil
 	}
 	return o.AccountID
 }
 
-func (o *EnhancedReportReportItem) GetAccountName() *string {
+func (o *ReportItem) GetAccountName() *string {
 	if o == nil {
 		return nil
 	}
 	return o.AccountName
 }
 
-func (o *EnhancedReportReportItem) GetBalance() *decimal.Big {
+func (o *ReportItem) GetBalance() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.Balance
 }
 
-func (o *EnhancedReportReportItem) GetDate() *string {
+func (o *ReportItem) GetDate() *string {
 	if o == nil {
 		return nil
 	}
@@ -86,7 +86,7 @@ func (o *EnhancedReportReportItem) GetDate() *string {
 type EnhancedReport struct {
 	ReportInfo *EnhancedReportInfo `json:"reportInfo,omitempty"`
 	// An array of report items.
-	ReportItems []EnhancedReportReportItem `json:"reportItems,omitempty"`
+	ReportItems []ReportItem `json:"reportItems,omitempty"`
 }
 
 func (o *EnhancedReport) GetReportInfo() *EnhancedReportInfo {
@@ -96,7 +96,7 @@ func (o *EnhancedReport) GetReportInfo() *EnhancedReportInfo {
 	return o.ReportInfo
 }
 
-func (o *EnhancedReport) GetReportItems() []EnhancedReportReportItem {
+func (o *EnhancedReport) GetReportItems() []ReportItem {
 	if o == nil {
 		return nil
 	}