+++ previous-versions/assess/pkg/models/shared/reportitems.go
@@ -9,21 +9,21 @@ import (
 	"github.com/ericlagergren/decimal"
 )
 
-// ReportItemsLoanTransactionType - The type of loan transaction.
-type ReportItemsLoanTransactionType string
+// LoanTransactionType - The type of loan transaction.
+type LoanTransactionType string
 
 const (
-	ReportItemsLoanTransactionTypeInvestment      ReportItemsLoanTransactionType = "Investment"
-	ReportItemsLoanTransactionTypeRepayment       ReportItemsLoanTransactionType = "Repayment"
-	ReportItemsLoanTransactionTypeInterest        ReportItemsLoanTransactionType = "Interest"
-	ReportItemsLoanTransactionTypeAccuredInterest ReportItemsLoanTransactionType = "AccuredInterest"
+	LoanTransactionTypeInvestment      LoanTransactionType = "Investment"
+	LoanTransactionTypeRepayment       LoanTransactionType = "Repayment"
+	LoanTransactionTypeInterest        LoanTransactionType = "Interest"
+	LoanTransactionTypeAccuredInterest LoanTransactionType = "AccuredInterest"
 )
 
-func (e ReportItemsLoanTransactionType) ToPointer() *ReportItemsLoanTransactionType {
+func (e LoanTransactionType) ToPointer() *LoanTransactionType {
 	return &e
 }
 
-func (e *ReportItemsLoanTransactionType) UnmarshalJSON(data []byte) error {
+func (e *LoanTransactionType) UnmarshalJSON(data []byte) error {
 	var v string
 	if err := json.Unmarshal(data, &v); err != nil {
 		return err
@@ -36,10 +36,10 @@ func (e *ReportItemsLoanTransactionType) UnmarshalJSON(data []byte) error {
 	case "Interest":
 		fallthrough
 	case "AccuredInterest":
-		*e = ReportItemsLoanTransactionType(v)
+		*e = LoanTransactionType(v)
 		return nil
 	default:
-		return fmt.Errorf("invalid value for ReportItemsLoanTransactionType: %v", v)
+		return fmt.Errorf("invalid value for LoanTransactionType: %v", v)
 	}
 }
 
@@ -71,7 +71,7 @@ type ReportItems struct {
 	LenderName *string  `json:"lenderName,omitempty"`
 	LoanRef    *LoanRef `json:"loanRef,omitempty"`
 	// The type of loan transaction.
-	LoanTransactionType *ReportItemsLoanTransactionType `json:"loanTransactionType,omitempty"`
+	LoanTransactionType *LoanTransactionType `json:"loanTransactionType,omitempty"`
 }
 
 func (r ReportItems) MarshalJSON() ([]byte, error) {
@@ -120,7 +120,7 @@ func (o *ReportItems) GetLoanRef() *LoanRef {
 	return o.LoanRef
 }
 
-func (o *ReportItems) GetLoanTransactionType() *ReportItemsLoanTransactionType {
+func (o *ReportItems) GetLoanTransactionType() *LoanTransactionType {
 	if o == nil {
 		return nil
 	}